langsmith_export/pyproject.toml
[tool.poetry]
name = "lang-smith-export"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langsmith = "^0.1.85"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
tqdm = "^4.66.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
